+++ ExpensesTracker/Shared/Models/User.cs
@@ -114,7 +114,7 @@ namespace ExpensesTracker.Shared.Models
                 sheet.Add(record);
             }
 
-            sheet = ListSort<Record>(sheet, "date", descending);
+            sheet = Sort.ListSort<Record>(sheet, "date", descending);
             return sheet;
         }
 
@@ -142,27 +142,5 @@ namespace ExpensesTracker.Shared.Models
             return recentIncome;
         }
 
-        public static List<T> ListSort<T>(List<T> data, string sortExpression, bool descending = false)
-        {
-            List<T>? sortedList = new List<T>();
-            if (!descending)
-            {
-                sortedList = (from n in data
-                              orderby GetSortProperty(n, sortExpression) ascending
-                              select n).ToList();
-            }
-            else 
-            {
-                sortedList = (from n in data
-                              orderby GetSortProperty(n, sortExpression) descending
-                              select n).ToList();
-            }
-            return sortedList;
-        }
-
-        public static object GetSortProperty(object item, string propName)
-        {
-            return item.GetType().GetProperty(propName).GetValue(item, null);
-        }
     }
 }
